+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [UNRELEASE]
 
+### Fixed 
+
+- Rename options related to ticket cloning and closure to avoid ambiguity
+
 ## [2.9.16] - 2025-07-10
 
 ### Fixed